feat: run all experiments at the same time
