+++ packages/simulator/data/parameters.json
@@ -8,13 +8,6 @@
     "bookings",
     "measureStations"
   ],
-  "mapInitState": {
-    "latitude": 55.901021,
-    "longitude": 13.001441,
-    "bearing": 0,
-    "zoom": 9,
-    "pitch": 40
-  },
   "fleets": {
     "Arjeplogs kommun": {
       "fleets": [
